requirements.txt
fastapi
bson
pymysql
fastapi-redis-cache
request
aiomysql
python-dotenv
pathlib
pyjwt
passlib
bcrypt
pydantic